Tesse-gym redesign for rllib / gnn work  (#81)
* add scene graph to observation and decode in policy

* resolve contiguous memory error

* add gcn definition; pass graph through network

* track current scene; check for falling agents

* pad graphs for batched observations

* add gcn support

* isort

* add config options for scene graphs

* example goseek scene graph config

* create agent that only observed graph

* fix typo

* turn off collisions

* WIP; expand options for graph only agent

* WIP; some minor gcn / logging updates

* batched graph inference

* add comments/typehints

* add option to query/use image data

* add missing args

* fix merge

* fix some config names

* initial code for scene graph loading

* add semantic label to node feature

* add feature to make targets invisible

* update config options'

* adjustable rnn input features

* added found targets to video logging

* pad with -1

* rotate graph points to agent frame. optimize scene graph creation

* record episode trajectory. formatting.

* rollback network updates

* fix typo

* fix transform from global to spawn point frame

* log visited nodes

* move step counting to core

* only mark place nodes as visited

* option to give node position relative to current agent pose. option to add linear layers to gcn

* make gcn out features final embedding size

* fix typo

* log trajectories

* call video logging at start of reset

* record scene in video logging

* get current scene from gym object

* fix scene graph to current agent frame transform

* option for esdfs

* add option for esdfs

* toggle modalities

* add option for configurable dense layers

* add cnn actor critic support for gcns

* add dsg compatible nav task

* add support for polar coords

* merge

* comments

* add coordinate option

* remove deprecated files

* remove unused configs

* refactoring

* add waypoint planner

* remove unnused files

* add agent frame support; refactoring

* remove print statements

* increase edge padding buffer

* reformatting

* add agent frame param config

* remove agent frame default args

* add edge threshold

* add sector frame

* update agent frame options

* removed unused code

* ESDF node definition; docs/refactoring

* add some gcn tests

* move gcn networks to own file

* sampled poses gcn + required refactoring; WIP

* reformatting

* reformatting

* reformatting

* error handing for logging videos

* remove v100 workaround

* add rank logic to training script

* add sim_ok argument

* add empty graph check

* rllib specific evaluation

* add docs, reformatting

* reformatting, docs

* fix episode length check

* video logging; give specific eval config

* add eval_name

* add random seed settings

* track scene changes per episode

* initial esdf freespace checking in agent layer

* change default edsf distance threshold

* update evaluation results save path

* add option to restore ckpt

* option to vary esdf slice level

* add back method

* fix typo

* Feature/esdf optimization (#73)

* esdf edge evaluation on gpu + in a batch

* allow for variable env args

* Feature/graph accumulation (#77)

* accumulate graph w/ depth sensor model

* fix bug in config parsing

* Feature/agent layer node type check (#76)

* option to specify node types

* arg name change

* adding agent layer config options

* arg name fix

* Fix/target track (#74)

* update target removal logic

* metadata notetype checking

* Feature/space tracking (#75)

* track explored space (and WIP error checking / esdf baseline features)

* log space visited

* fix typo

* support for multiple evaluation workers

* support for eval config partition with rank offset

* option to use esdf raycast for agent layer checks

* Feature/esdf obs (#79)

* initial esdf observation (WIP)

* initial esdf obs feature

* esdf slice feature

* log custom stats (#78)

* optimize esdf observation forming

* comments

* comments

* rename init function; add todo

* add notebook to run agent

* round of refactoring to gcn code

* round of refactoring to networks

* renaming

* add docs, some refactoring

* simpliy scene graph task

* reorganize code

* refactor

* rename private vars

* refactoring

* cleanup

* clean up action layer

* rename to action layer

* remove unused args

* refactor; allow for flexible action and observation spaces

* update observation space handling

* scene graph observer default args

* base observer

* round of refactoring

* remove unused test

* gitignore

* save notebooks as jupytext

* remove rllib specific code

* update gitignore

* remove dependency on scene_graph_learning

* remove notebook

* refactored load_graph method

* make compatible with vector spaces

* move training scripts out of tesse-gym

* logic moved to dsg_rl

* remove scripts

* move action mapper to own file

* update readme

* update observer name

* update link

+++ src/tesse_gym/core/observations.py
@@ -25,7 +25,7 @@ from typing import Dict, List, Optional, Tuple, Union
 import gym.spaces as spaces
 import numpy as np
 
-from tesse.msgs import Camera, Channels, Compression, DataResponse
+from tesse.msgs import Camera, Channels, Compression
 
 ObservationConfig = namedtuple(
     "ObservationConfig",
@@ -37,15 +37,15 @@ ObservationConfig = namedtuple(
 def setup_observations(
     observation_config: ObservationConfig,
 ) -> Tuple[Tuple[Camera, Compression, Channels], Union[spaces.Box, spaces.Dict]]:
-    """ Creates observation configuration for `TesseGym`.
+    """Creates observation configuration for `TesseGym`.
 
     Parameters:
-        observation_config (ObservationConfig): Object 
+        observation_config (ObservationConfig): Object
             containing info to specify gym observations.
-    
+
     Returns:
         Tuple[Tuple[Camera, Compression, Channels], spaces.Box]
-            - Tuple listing TESSE cameras to query when getting an 
+            - Tuple listing TESSE cameras to query when getting an
                 observation.
             - spaces.Box object describing the observation space.
     """
@@ -81,21 +81,21 @@ def get_observation_space(
     use_dict: Optional[bool] = True,
     custom_obs: Optional[Dict[str, Tuple[float, float, Tuple[int, ...]]]] = None,
 ) -> Union[spaces.Box, spaces.Dict]:
-    """ Get observation space.
-    
+    """Get observation space.
+
     Parameters:
-        observation_modalities (Tuple[Camera, ...]): Tuple of observation 
+        observation_modalities (Tuple[Camera, ...]): Tuple of observation
             modalities (e.g., RGB_LEFT, DEPTH, SEGMENTATION, ...)
         height_in_pixels (int): Observation image height.
         width_in_pixels (int): Observation image width.
         pose (bool): True if pose is included in observation.
         obs_min (float): Min possible observation value.
-        obs_max (float): Max possible observation value. 
+        obs_max (float): Max possible observation value.
         use_dict (bool): Use dict space. Otherwise, flatten
             observations into vector.
-    
+
     Returns:
-        spaces.Box: OpenAI Gym spaces.Box object describing the 
+        spaces.Box: OpenAI Gym spaces.Box object describing the
             observation given as arguments.
     """
     obs_names = []
@@ -106,12 +106,19 @@ def get_observation_space(
 
     if use_dict:
         obs_spaces = [
-            spaces.Box(obs_min, obs_max, shape=(height_in_pixels, width_in_pixels, c))
+            spaces.Box(
+                obs_min,
+                obs_max,
+                shape=(height_in_pixels, width_in_pixels, c),
+                dtype=np.float64,
+            )
             for c in camera_channels
         ]
         if pose:
             obs_names.append("POSE")
-            obs_spaces.append(spaces.Box(obs_min, obs_max, shape=(3,)))
+            obs_spaces.append(
+                spaces.Box(obs_min, obs_max, shape=(3,), dtype=np.float64)
+            )
 
         if custom_obs is not None:
             for k, (custom_min, custom_max, custom_shape) in custom_obs.items():
@@ -134,20 +141,20 @@ def decode_observations(
     img_channels: Tuple[int, ...] = (3, 1, 1),
     pose_shapes: Tuple[int, ...] = (3,),
 ) -> Tuple[np.ndarray, ...]:
-    """ Decode observation vector into images and poses.
+    """Decode observation vector into images and poses.
 
     Args:
         observation (np.ndarray): Shape (N,) observation array of flattened
             images concatenated with a pose vector. Thus, N is equal to N*H*W*C + N*3.
         img_shape (Tuple[int, int]): (H, W) of all images.
         img_channels (Tuple[int, ...]): Number of channles of each image in `observation`.
-        pose_shapes (Tuple[int, ...]): Shape of all poses stacked 
+        pose_shapes (Tuple[int, ...]): Shape of all poses stacked
 
     Returns:
         Tuple[np.ndarray, ...]: Arrays of images and poses.
 
     Notes:
-        Assumes images are flattened and stacked first, followed by poses. 
+        Assumes images are flattened and stacked first, followed by poses.
     """
     decoded_obs = []
